conversion to original cppunit has been done

+++ Foundation/testsuite/src/TextConverterTest.cpp
@@ -39,13 +39,13 @@ void TextConverterTest::testIdentityASCII()
 {
 	ASCIIEncoding encoding;
 	TextConverter converter(encoding, encoding);
-	
+
 	std::string empty;
 	std::string result0;
 	int errors = converter.convert(empty, result0);
 	assert (result0 == empty);
 	assert (errors == 0);
-	
+
 	std::string fooBar = "foo bar";
 	std::string result1;
 	errors = converter.convert(fooBar, result1);
@@ -61,7 +61,7 @@ void TextConverterTest::testIdentityASCII()
 	errors = converter.convert("", 0, result3);
 	assert (result3.empty());
 	assert (errors == 0);
-	
+
 	std::string x = "x";
 	std::string result4;
 	errors = converter.convert(x, result4);
@@ -79,13 +79,13 @@ void TextConverterTest::testIdentityUTF8()
 {
 	UTF8Encoding encoding;
 	TextConverter converter(encoding, encoding);
-	
+
 	std::string empty;
 	std::string result0;
 	int errors = converter.convert(empty, result0);
 	assert (result0 == empty);
 	assert (errors == 0);
-	
+
 	std::string fooBar = "foo bar";
 	std::string result1;
 	errors = converter.convert(fooBar, result1);
@@ -101,20 +101,20 @@ void TextConverterTest::testIdentityUTF8()
 	errors = converter.convert("", 0, result3);
 	assert (result3.empty());
 	assert (errors == 0);
-	
+
 	const unsigned char greek[] = {0x20, 0xce, 0xba, 0xe1, 0xbd, 0xb9, 0xcf, 0x83, 0xce, 0xbc, 0xce, 0xb5, 0x20, 0x00};
 	std::string text((const char*) greek);
-	
+
 	std::string result4;
 	errors = converter.convert(text, result4);
 	assert (result4 == text);
 	assert (errors == 0);
-	
+
 	std::string result5;
 	errors = converter.convert((char*) greek, 13, result5);
 	assert (result5 == text);
 	assert (errors == 0);
-	
+
 	std::string x = "x";
 	std::string result6;
 	errors = converter.convert(x, result6);
@@ -125,13 +125,13 @@ void TextConverterTest::testIdentityUTF8()
 	errors = converter.convert("x", 1, result7);
 	assert (result7 == x);
 	assert (errors == 0);
-	
+
 	std::string utfChar((char*) greek + 1, 2);
 	std::string result8;
 	errors = converter.convert(utfChar, result8);
 	assert (result8 == utfChar);
 	assert (errors == 0);
-	
+
 	std::string result9;
 	errors = converter.convert((char*) greek + 1, 2, result9);
 	assert (result9 == utfChar);
@@ -151,7 +151,7 @@ void TextConverterTest::testUTF8toASCII()
 	int errors = converter.convert(text, result0);
 	assert (result0 == " ????? AB");
 	assert (errors == 0);
-	
+
 	std::string result1;
 	errors = converter.convert("abcde", 5, result1);
 	assert (result1 == "abcde");
@@ -163,18 +163,18 @@ void TextConverterTest::testLatin1toUTF8()
 	Latin1Encoding latin1Encoding;
 	UTF8Encoding utf8Encoding;
 	TextConverter converter(latin1Encoding, utf8Encoding);
-	
+
 	const unsigned char latin1Chars[] = {'g', 252, 'n', 't', 'e', 'r', 0};
 	const unsigned char utf8Chars[]   = {'g', 195, 188, 'n', 't', 'e', 'r', 0};
 	std::string latin1Text((const char*) latin1Chars);
 	std::string utf8Text((const char*) utf8Chars);
-	
+
 	std::string result0;
 	int errors = converter.convert(latin1Text, result0);
 	assert (result0 == utf8Text);
 	assert (errors == 0);
-	assertEqual((long) result0.size(), 7);
-	
+	assertEqual( result0.size(), size_t( 7 ) );
+
 	std::string result1;
 	errors = converter.convert(latin1Chars, 6, result1);
 	assert (result1 == utf8Text);
@@ -198,13 +198,13 @@ void TextConverterTest::testLatin2toUTF8()
 	int errors = converter.convert(latinText, result0);
 	assertEqual (result0, utf8Text);
 	assertEqual (errors, 0);
-	assertEqual((long) result0.size(), 49);
+	assertEqual( result0.size(), size_t( 49 ) );
 
 	std::string result1;
 	errors = converter.convert(latinChars, 25, result1);
 	assertEqual (result1, utf8Text);
 	assertEqual (errors, 0);
-	assertEqual((long) result1.size(), 49);
+	assertEqual( result1.size(), size_t( 49 ) );
 }
 
 
@@ -224,13 +224,13 @@ void TextConverterTest::testLatin9toUTF8()
 	int errors = converter.convert(latinText, result0);
 	assertEqual (result0, utf8Text);
 	assertEqual (errors, 0);
-	assertEqual((long) result0.size(), 43);
+	assertEqual( result0.size(), size_t( 43 ) );
 
 	std::string result1;
 	errors = converter.convert(latinChars, 25, result1);
 	assertEqual(result1, utf8Text);
-	assertEqual((long) errors, 0);
-	assertEqual((long) result1.size(), 43);
+	assertEqual( errors, 0);
+	assertEqual( result1.size(), size_t( 43 ) );
 }
 
 
@@ -250,13 +250,13 @@ void TextConverterTest::testCP1250toUTF8()
 	int errors = converter.convert(latinText, result0);
 	assertEqual (result0, utf8Text);
 	assertEqual (errors, 0);
-	assertEqual((long) result0.size(), 49);
+	assertEqual( result0.size(), size_t( 49 ) );
 
 	std::string result1;
 	errors = converter.convert(latinChars, 25, result1);
 	assertEqual(result1, utf8Text);
-	assertEqual((long) errors, 0);
-	assertEqual((long) result1.size(), 49);
+	assertEqual( errors, 0 );
+	assertEqual( result1.size(), size_t( 49 ) );
 }
 
 
@@ -275,13 +275,13 @@ void TextConverterTest::testCP1251toUTF8()
 	int errors = converter.convert(latinText, result0);
 	assertEqual (result0, utf8Text);
 	assertEqual (errors, 0);
-	assertEqual((long) result0.size(), 62);
+	assertEqual( result0.size(), size_t( 62 ) );
 
 	std::string result1;
 	errors = converter.convert(latinChars, 31, result1);
 	assertEqual (result1, utf8Text);
 	assertEqual (errors, 0);
-	assertEqual((long) result1.size(), 62);
+	assertEqual( result1.size(), size_t( 62 ) );
 }
 
 
@@ -301,13 +301,13 @@ void TextConverterTest::testCP1252toUTF8()
 	int errors = converter.convert(latinText, result0);
 	assertEqual(result0, utf8Text);
 	assertEqual(errors, 0);
-	assertEqual((long) result0.size(), 43);
+	assertEqual( result0.size(), size_t( 43 ) );
 
 	std::string result1;
 	errors = converter.convert(latinChars, 25, result1);
 	assertEqual(result1, utf8Text);
 	assertEqual(errors, 0);
-	assertEqual((long) result1.size(), 43);
+	assertEqual( result1.size(), size_t( 43 ) );
 }
 
 
@@ -319,7 +319,7 @@ void TextConverterTest::testErrors()
 
 	const unsigned char badChars[] = {'a', 'b', 255, 'c', 254, 0};
 	std::string badText((const char*) badChars);
-	
+
 	std::string result;
 	int errors = converter.convert(badText, result);
 	assert (errors == 2);
